refactor: extract expense card action handler
